+++ tilings/strategies/__init__.py
@@ -25,6 +25,7 @@ from .requirement_insertion import (
     RequirementExtensionFactory,
     RequirementInsertionFactory,
     RootInsertionFactory,
+    TargetedCellInsertionFactory,
 )
 from .requirement_placement import (
     AllPlacementsFactory,
@@ -64,6 +65,7 @@ __all__ = [
     "RequirementCorroborationFactory",
     "RootInsertionFactory",
     "RowAndColumnPlacementFactory",
+    "TargetedCellInsertionFactory",
     # Decomposition
     "FactorFactory",
     # Deflation